fixed initial page load not working correctly. Now all that's left is hmr...
