+++ packages/website/package.json
@@ -15,12 +15,12 @@
   },
   "dependencies": {
     "@marko/express": "^1.0.0",
-    "@marko/tags-api-preview": "^0.2.0",
+    "@marko/tags-api-preview": "^0.2.1",
     "@mikro-ui/theme-default": "^1.0.0",
     "@mikro-ui/tokens": "^1.0.0",
     "compression": "^1.7.4",
     "express": "^4.17.2",
-    "marko": "^5.18.2",
+    "marko": "^5.19.0",
     "mikro-ui": "1.0.0",
     "remark-mikro": "1.0.0"
   },
@@ -30,11 +30,11 @@
   "devDependencies": {
     "@marko/vite": "^2.2.6",
     "@originjs/vite-plugin-commonjs": "^1.0.2",
-    "esbuild": "^0.14.14",
+    "esbuild": "^0.14.17",
     "remark-loader": "^4.0.0",
-    "sass": "^1.44.0",
-    "sass-loader": "^12.3.0",
-    "vite": "^2.7.10"
+    "sass": "^1.49.7",
+    "sass-loader": "^12.4.0",
+    "vite": "^2.7.13"
   },
   "volta": {
     "node": "16.13.2"
